chore: Return signatureMethod at file list
Used to be possible load SignPDF as internal page only using data from
file list.
